Load one or all database nodes for a server
With a given connection, it is now possible to specify the number of
database nodes for a server: Only one node with the explicit given
database can be loaded or all databases on the server. This is realized
by an extra parameter for the database connection in the tree model and
connection dialog.